bot/main.py
import os

import discord
from discord import app_commands
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

DISCORD_TOKEN = os.getenv("DISCORD_TOKEN")


class ArtifactOptimizerBot(discord.Client):
    """Discord bot for Genshin Impact artifact optimization."""

    def __init__(self):
        intents = discord.Intents.default()
        intents.message_content = True
        super().__init__(intents=intents)
        self.tree = app_commands.CommandTree(self)

    async def setup_hook(self):
        """Called when the bot is ready to sync commands."""
        await self.tree.sync()

    async def on_ready(self):
        """Called when the bot has connected to Discord."""
        print(f"Logged in as {self.user} (ID: {self.user.id})")
        print("------")


client = ArtifactOptimizerBot()


@client.tree.command(name="optimize", description="Recommend optimal artifacts for all 5 slots based on your inventory")
@app_commands.describe(character="The character to optimize artifacts for")
async def optimize(interaction: discord.Interaction, character: str):
    """
    Placeholder command for artifact optimization.

    This command will analyze the user's artifact inventory and recommend
    the optimal artifact for each of the 5 slots (Flower, Plume, Sands,
    Goblet, Circlet) based on calculated sub-stat efficiency.
    """
    # TODO: Implement artifact optimization logic
    # 1. Fetch user's inventory from database
    # 2. Calculate sub-stat efficiency for each artifact
    # 3. Find optimal artifact for each of the 5 slots for the specified character
    # 4. Return recommendations to user

    await interaction.response.send_message(
        f"🔧 Artifact optimization for **{character}** is coming soon!\n\n"
        "This feature will analyze your artifact inventory and recommend "
        "the optimal artifact for each of the 5 slots based on sub-stat efficiency.",
        ephemeral=True
    )


def main():
    """Entry point for the bot."""
    if not DISCORD_TOKEN:
        raise ValueError("DISCORD_TOKEN environment variable is not set. Check your .env file.")
    client.run(DISCORD_TOKEN)


if __name__ == "__main__":
    main()
